convert to u32

+++ src/widgets/grid_view.rs
@@ -17,7 +17,7 @@ impl IWidget for GridView {
         // rectangle(WHITE, rect, ctx.transform, gl);
     }
 
-    fn layout(&mut self, _width: f64, _height: f64) {
+    fn layout(&mut self, _width: u32, _height: u32) {
     }
 
     fn get_width(&self) -> Size {
